Complete design rewrite.
Major changes:

* Rewrite from C++ to C.
Since new design didn't require any features of C++, I decided it would
be better off with C linkage, since integration with other languages
will be much easier this way. The tests are still written in C++ though.

* Add decoder hooks
The previous implementation used C++ inheritance to have access to all
of the supported decoders. Since that actually introduced some runtime
overhead (we still need to cast each decoder to correct type, for known
compile time offsets) and to satisfy the decision of language change,
I decided to use decoder hooks to keep the ability to have separated
decoders each for their own ISA. Also this way we can keep the previous
implementation of each decoder with some minor changes.

The actual need for separate decoders is up for a debate. On the first
hand, it really smoothes the implementation with the ability to separate
ISA decoders. On the other hand, it introduces tiny overhead of having
to loop over each decoder.